requirements.txt
openai
httpx
python-dotenv
flask
